backend/pyproject.toml
[project]
name = "ouroboros-backend"
version = "0.1.0"
description = "Backend services for Ouroboros"
requires-python = ">=3.11"
dependencies = [
  "fastapi==0.115.8",
  "pydantic-settings==2.8.1",
  "uvicorn[standard]==0.34.0",
  "biopython==1.84"
]

[project.optional-dependencies]
dev = [
  "httpx==0.28.1",
  "pytest==8.3.5"
]

[tool.pytest.ini_options]
pythonpath = ["."]
